Add viewWillAppear testing
