+++ GuildedRoseLLC/ItemCollectionViewDataSource.swift
@@ -2,7 +2,7 @@ import UIKit
 
 public class ItemCollectionViewDataSource: NSObject, UICollectionViewDataSource {
     
-    var items: [Item] = []
+    public var items: [Item] = []
 
     public func show(items: [Item]) {
         self.items = items
